Apply bugfixes and update function signatures for cross-platform build
- Add /DBUGFIX to NOASM build flags in Makefile
- Update function signatures in headers and stubs for consistency
- Remove MSDOS-specific #ifdefs in favor of BUGFIX where needed
- Fix types and prototypes for openFile, createFile, showPicFile, etc.
- Add missing includes and externs for cross-module references
- Rename clearKeyBuf to clearKeybuf for consistency
- Update util and util2 functions to match corrected signatures
- Adjust slot.h and gfx_impl.c for updated overlay slot prototypes
- Fix inttype.h typedefs and includes for portability
- Remove redundant or incorrect extern/global variable definitions
- Ensure all function declarations and calls match across modules

+++ src/stdata.c
@@ -2,6 +2,7 @@
 #include "struct.h"
 #include "comm.h"
 #include <stdio.h>
+#include <dos.h>
 
 /* === Group 1 (0x0042-0x0530): Filename, mission selection, UI strings === */
 char aLabs_pic[] = "labs.pic";
@@ -706,7 +707,7 @@ char todayMissStrBuf[0x1D];
 uint8 missionStrTrunc;
 uint8 missionStrTruncEnd;
 uint8 exitCode[2];
-int *nearestTerrainResult;
+int16 *nearestTerrainResult;
 char objectTypeTable[0x64];
 uint8 wldReadBuf8[0x64];
 uint8 wldReadBuf7[0x64];
